The initial motivation behind basictests was for a quick set of sanity tests to check whether your code made sense. It was not meant for thorough testing to find corner cases. However, over time, it grew to include such expensive tests. This PR takes them out. It also allows us to bring back basictests to MinGW tests, since it is now cheap.
This is not an exercise in software engineering and making things prettier. This is a pragmatic change to improve our
test coverage and quality of life.
